requirements.txt
langchain==0.3.16
langchain-community==0.3.16
langchain-google-genai==2.0.8
langchain-core==0.3.33

python-dotenv==1.0.1
faiss-cpu==1.12.0
pypdf==5.1.0
pydantic-settings==2.6.1
loguru==0.7.3
rich==13.9.4

pytest==8.3.4
ruff==0.8.4
mypy==1.13.0
pre-commit==4.0.1
